:hammer: restructure openvas parser (#9658)
* :hammer: restructure openvas parser

* adapt csv parser

* fix csv parser

* fix xml parser

* flake8
